% file MASS/man/chis2009.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{chis2009}
\alias{chis2009}
\title{
  CHIS 2009 Adult Health Survey Data
}
\description{
  The \code{chis2009} data frame has 17179 rows and 26 columns.
}
\usage{
chis2009
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{\code{ab1}}{
      GENERAL HEALTH CONDITION
    }
    \item{\code{ac13}}{
     NUMBER OF TIMES DRANK FRUIT-FLAV LAST MONTH - UNIT
    }
    \item{\code{ac14}}{
     NUMBER OF TIMES ATE ICE CREAM/FROZEN DESSERTS LAST MONTH
    }
    \item{\code{ad41w}}{
     NUMBER OF TIMES WALKED AT LEAST 10 MIN FOR LEISURE PAST 7 DAYS     
    }
    \item{\code{ad42w}}{
	AVERAGE LENGTH OF TIME WALKED FOR LEISURE
    }
    \item{\code{ae2}}{
     NUMBER OF TIMES ATE FRUIT IN PAST MO
    }
    \item{\code{ae27}}{
     NUMBER OF DAYS MODERATE PHYSICAL ACTIVITY IN PAST WEEK
    }
    \item{\code{ae27a}}{
	TIME PER DAY OF MODERATE PHYSICAL ACTIVITY
    }
    \item{\code{ae3}}{
     NUMBER OF TIMES ATE FRNCH FRIES, HME FRIES, HSH BRWNS IN PAST
MO
    }
    \item{\code{ae7}}{
     NUMBER OF TIMES ATE VEGETABLES IN PAST MO
    }
    \item{\code{ah5}}{
     NUMBER OF TIMES SAW MD IN PAST 12 MOS
    }
    \item{\code{ak3}}{
     NUMBER OF  USUAL HRS WORKED PER WEEK
    }
    \item{\code{ak7}}{
	LENGTH OF TIME WORKING AT MAIN JOB
    }
    \item{\code{distress}}{
SERIOUS PSYCHOLOGICAL DISTRESS
    }
    \item{\code{aheduc}}{
	EDUCATIONAL ATTAINMENT
    }
    \item{\code{timead}}{
	LENGTH OF TIME LIVED AT CURRENT ADDRESS (IN MONTHS)
    }
    \item{\code{ak10_p}}{
	RESPONDENT'S EARNINGS LAST MONTH
    }
    \item{\code{ak22_p}}{
	HOUSEHOLD'S TOTAL ANNUAL INC
    }
    \item{\code{heighm_p}}{
	HEIGHT: METERS
    }
    \item{\code{srage_p}}{
	AGE
    }
    \item{\code{wt18k_p}}{
	WEIGHT AT 18: KILOS
    }
    \item{\code{sug_past}}{
	UNADJUSTED DAILY TEASPOONS OF ADDED SUGAR IN PASTRIES
    }
    \item{\code{sug_bev}}{
	UNADJUSTED DAILY TEASPOONS OF ADDED SUGAR IN ALL
	BEVERAGES
    }
    \item{\code{fv_nobns}}{
	DAILY CUP EQUIVALENTS OF FRUITS AND VEGETABLES EXCLUDING
	BEANS
    }
    \item{\code{sugar2}}{
	DAILY TEASPOONS OF ADDED SUGAR
    }
    \item{\code{Weight}}{
		WEIGHT: KG
    }
  }
}
\details{
 The 2009 California Health Interview Survey
(CHIS 2009). The CHIS is a population based telephone survey of California's population.
The survey aims to collect extensive information on health status, health conditions, health 
related behaviors, health insurance coverage as well as access to health care services. Within
each household, separate interviews are conducted with a randomly selected adult (age 18
and over). The dataset consists of 536 features measured for 47614 respondents.

}
\source{
 CHIS California Health Interview Survey. Los Angeles (CA). 
UCLA Center for Health Policy Research. \url{http://www.chis.ucla.edu/}.  
}
\keyword{datasets}
