\name{DetR-package}
\alias{DetR-package}
\docType{package}
\title{ Deterministic and Robust Algorithms for Regression.}
\description{
This packages contains various robust and deterministic algorithms
 for linear regression.
}
\details{
\tabular{ll}{
Package: \tab DetR\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-09-19\cr
Depends: \tab matrixStats, robustbase, MASS\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
DetR-package           Robust and Deterministic Algothms for Linear 
                        Regression
DetLTS                  DetLTS algorithm (deterministic counterpart of FastLTS).
OGKCStep               Tests of OGK+Csteps.
DetMM                  DetMM algorithm (deterministic counterpart of FastMM).
test_function		unit test functions.
}
}
\author{
Kaveh Vakili [aut, cre], using translation and modifications of codes
 from other packages (see Desrciption and the individual fuctions' helpfiles)

Maintainer: Kaveh Vakili <vakili.kaveh.email@gmail.com>
}
\keyword{package}
